refactoring of cosmos_client to be private for imports
