+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/__init__.py
@@ -20,7 +20,7 @@
 # SOFTWARE.
 
 from ._container import ContainerProxy
-from .cosmos_client import CosmosClient
+from ._cosmos_client import CosmosClient
 from ._database import DatabaseProxy
 from ._user import UserProxy
 from ._scripts import ScriptsProxy
