+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/_cosmos_client.py
